+++ packages/terra-date-time-picker/tests/jest/TimeClarification.test.jsx
@@ -36,9 +36,10 @@ it('should render a disabled time clarification', () => {
 });
 
 it('should render offset button after daylight savings button clicked', () => {
+  moment.tz.guess = jest.fn(() => 'America/Chicago');
   const datePicker = shallowWithIntl((
     <TimeClarification
-      ambiguousDateTime={moment.tz('2019-08-15T10:30:00', 'America/Chicago')}
+      ambiguousDateTime="2019-08-15T10:30:00"
       isOpen
       isOffsetButtonHidden
       onDaylightSavingButtonClick={jest.fn()}
@@ -54,9 +55,10 @@ it('should render offset button after daylight savings button clicked', () => {
 });
 
 it('should render offset button after standard time button clicked', () => {
+  moment.tz.guess = jest.fn(() => 'America/Chicago');
   const datePicker = shallowWithIntl((
     <TimeClarification
-      ambiguousDateTime={moment.tz('2019-02-15T10:30:00', 'America/Chicago')}
+      ambiguousDateTime="2019-02-15T10:30:00"
       isOpen
       isOffsetButtonHidden
       onDaylightSavingButtonClick={jest.fn()}
@@ -72,9 +74,10 @@ it('should render offset button after standard time button clicked', () => {
 });
 
 it('should render offset button after daylight savings button clicked in the southern hemisphere timezone', () => {
+  moment.tz.guess = jest.fn(() => 'Australia/Sydney');
   const datePicker = shallowWithIntl((
     <TimeClarification
-      ambiguousDateTime={moment.tz('2019-04-01T02:30:00', 'Australia/Sydney')}
+      ambiguousDateTime="2019-04-01T02:30:00"
       isOpen
       isOffsetButtonHidden
       onDaylightSavingButtonClick={jest.fn()}
@@ -90,9 +93,10 @@ it('should render offset button after daylight savings button clicked in the sou
 });
 
 it('should render offset button after standard time button clicked in the southern hemisphere timezone', () => {
+  moment.tz.guess = jest.fn(() => 'Australia/Sydney');
   const datePicker = shallowWithIntl((
     <TimeClarification
-      ambiguousDateTime={moment.tz('2019-10-01T02:30:00', 'Australia/Sydney')}
+      ambiguousDateTime="2019-10-01T02:30:00"
       isOpen
       isOffsetButtonHidden
       onDaylightSavingButtonClick={jest.fn()}